# src/main/webui/openapi.json
{
  "openapi": "3.1.0",
  "components": {
    "schemas": {
...
      "FailedComponent": {
        "description": "Metadata of submitted components failed to be processed for scanning",
        "type": "object",
        "required": [
          "name",
          "version",
          "image",
          "error"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Component name"
          },
          "version": {
            "type": "string",
            "description": "Component version"
          },
          "image": {
            "type": "string",
            "description": "Component image"
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "ProductSummary": {
        "description": "Product metadata and reports data",
        "type": "object",
        "required": [
          "data",
          "summary"
        ],
        "properties": {
          "data": {
            "description": "Product data",
            "type": "object",
            "required": [
              "id",
              "name",
              "version",
              "submittedAt",
              "submittedCount",
              "metadata",
              "submissionFailures",
              "cveId"
            ],
            "properties": {
              "id": {
                "type": "string",
                "description": "Product ID"
              },
              "name": {
                "type": "string",
                "description": "Product name"
              },
              "version": {
                "type": "string",
                "description": "Product version"
              },
              "submittedAt": {
                "type": "string",
                "description": "Timestamp of product scan request submission"
              },
              "submittedCount": {
                "type": "integer",
                "format": "int32",
                "description": "Number of components submitted for scanning"
              },
              "metadata": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "description": "Product user provided metadata"
              },
              "submissionFailures": {
                "description": "List of submitted components failed to be processed for scanning",
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FailedComponent"
                }
              },
              "completedAt": {
                "type": "string",
                "description": "Timestamp of product scan request completion"
              },
              "cveId": {
                "type": "string",
                "description": "CVE ID associated with this product"
              }
            }
          },
          "summary": {
            "description": "Product reports summary data",
            "type": "object",
            "required": [
              "productState",
              "statusCounts",
              "justificationStatusCounts"
            ],
            "properties": {
              "productState": {
                "type": "string",
                "description": "Product state of analysis"
              },
              "statusCounts": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "format": "int32"
                },
                "description": "Map of analysis state to count of reports with that state"
              },
              "justificationStatusCounts": {
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "format": "int32"
                },
                "description": "Map of justification status to count of reports with that status"
              }
            }
          }
        }
      },
...
      "ReportData": {
        "description": "A single report request",
        "type": "object",
        "required": [
          "reportRequestId",
          "report"
        ],
        "properties": {
          "reportRequestId": {
            "description": "Report identification",
            "type": "object",
            "required": [
              "id",
              "reportId"
            ],
            "properties": {
              "id": {
                "type": "string",
                "description": "ID assigned for database identification"
              },
              "reportId": {
                "type": "string",
                "description": "ID assigned on scan request submission"
              }
            }
          },
          "report": {
            "description": "Report payload",
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "description": "Please enter your JWT Token without Bearer",
        "scheme": "bearer",
        "bearerFormat": "jwt"
      },
      "SecurityScheme": {
        "type": "openIdConnect",
        "description": "Authentication"
      }
    }
  },
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "summary": "List all product data",
        "description": "Retrieves paginated, sortable, and filterable product data for all products",
        "parameters": [
          {
            "name": "cveId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": "0"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": "100"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "DESC"
            }
          },
          {
            "name": "sortField",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "submittedAt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductSummary"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Product Endpoint"
        ]
      }
    },
    "/api/v1/products/upload-cyclonedx": {
      "post": {
        "summary": "Upload CycloneDX file for analysis",
        "description": "Accepts a CVE ID and CycloneDX JSON file, validates them, and queues the report for analysis",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "cveId": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "File uploaded and queued for analysis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportData"
                }
              }
            }
          },
          "400": {
            "description": "Validation error with field-specific error messages"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Product Endpoint"
        ]
      }
    },
    "/api/v1/products/upload-spdx": {
      "post": {
        "summary": "Create new product from SPDX SBOM",
        "description": "Uploads an SPDX SBOM file, parses it, creates a product, and starts async processing. Requires a vulnerability ID to include in all component reports.",
        "parameters": [
          {
            "description": "Vulnerability ID (e.g., CVE-2024-12345) to include in all component reports",
            "required": true,
            "name": "vulnerabilityId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Product creation request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid SPDX file, missing required data, or missing vulnerability ID"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Product Endpoint"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "delete": {
        "parameters": [
          {
            "description": "Product ID",
            "required": true,
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Product deletion request accepted"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Remove",
        "tags": [
          "Product Endpoint"
        ]
      },
      "get": {
        "summary": "Get product",
        "description": "Gets product by ID from database",
        "parameters": [
          {
            "description": "Product ID",
            "required": true,
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found in database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSummary"
                }
              }
            }
          },
          "404": {
            "description": "Product not found in database"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Product Endpoint"
        ]
      }
    },
...
    "/api/v1/reports/product/{id}": {
      "delete": {
        "summary": "Delete product by ID",
        "description": "Deletes all component analysis reports and product metadata associated with a specific product ID",
        "parameters": [
          {
            "description": "Product ID to delete",
            "required": true,
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Product deletion request accepted"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      },
      "get": {
        "summary": "Get product data by ID",
        "description": "Retrieves product data for a specific product ID",
        "parameters": [
          {
            "description": "Product ID",
            "required": true,
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Product metadata and reports data",
                  "type": "object",
                  "required": [
                    "data",
                    "summary"
                  ],
                  "properties": {
                    "data": {
                      "description": "Product data",
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "version",
                        "submittedAt",
                        "submittedCount",
                        "metadata",
                        "submissionFailures",
                        "cveId"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "Product ID"
                        },
                        "name": {
                          "type": "string",
                          "description": "Product name"
                        },
                        "version": {
                          "type": "string",
                          "description": "Product version"
                        },
                        "submittedAt": {
                          "type": "string",
                          "description": "Timestamp of product scan request submission"
                        },
                        "submittedCount": {
                          "type": "integer",
                          "format": "int32",
                          "description": "Number of components submitted for scanning"
                        },
                        "metadata": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          },
                          "description": "Product user provided metadata"
                        },
                        "submissionFailures": {
                          "description": "List of submitted components failed to be processed for scanning",
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FailedComponent"
                          }
                        },
                        "completedAt": {
                          "type": "string",
                          "description": "Timestamp of product scan request completion"
                        },
                        "cveId": {
                          "type": "string",
                          "description": "CVE ID associated with this product"
                        }
                      }
                    },
                    "summary": {
                      "description": "Product reports summary data",
                      "type": "object",
                      "required": [
                        "productState",
                        "statusCounts",
                        "justificationStatusCounts"
                      ],
                      "properties": {
                        "productState": {
                          "type": "string",
                          "description": "Product state of analysis"
                        },
                        "statusCounts": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "description": "Map of analysis state to count of reports with that state"
                        },
                        "justificationStatusCounts": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "description": "Map of justification status to count of reports with that status"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      }
    },
...
  },
...
  "info": {
    "title": "exploit-iq-client API",
    "version": "1.0.0-SNAPSHOT"
  }
}